use fixture to check that stack is deployed before testing
