+++ tests/swarm-deploy/test_service_restart.py
@@ -3,78 +3,14 @@
 # pylint: disable=unused-variable
 
 import logging
-import subprocess
 import time
 from pprint import pformat
-from typing import Dict, List
+from typing import List
 
 import pytest
-from docker import DockerClient
 from docker.models.services import Service
-from tenacity import Retrying, before_log, stop_after_attempt, wait_fixed
 
 log = logging.getLogger(__name__)
-
-
-# time measured from command 'up' finished until *all* tasks are running
-MAX_TIME_TO_DEPLOY_SECS = 60
-MAX_TIME_TO_RESTART_SERVICE = 10
-
-
-@pytest.fixture(scope="module")
-def deployed_simcore_stack(
-    core_stack_name: str, core_stack_compose: Dict, docker_client: DockerClient
-) -> List[Service]:
-
-    # NOTE: the goal here is NOT to test time-to-deplopy but
-    # rather guaranteing that the framework is fully deployed before starting
-    # tests. Obviously in a critical state in which the frameworks has a problem
-    # the fixture will fail
-
-    try:
-        for attempt in Retrying(
-            wait=wait_fixed(MAX_TIME_TO_DEPLOY_SECS),
-            stop=stop_after_attempt(5),
-            before=before_log(log, logging.WARNING),
-        ):
-            with attempt:
-                for service in docker_client.services.list():
-                    for task in service.tasks():
-                        # NOTE: Could have been restarted from latest test parameter, accept as well complete
-                        assert task["Status"]["State"] in (
-                            task["DesiredState"],
-                            "complete",
-                        ), (
-                            f"{service.name} still not ready or complete. Expected "
-                            f"desired_state[{task['DesiredState']}] but got "
-                            f"status_state[{task['Status']['State']}]). Details:"
-                            f"\n{pformat(task)}"
-                        )
-
-    finally:
-        subprocess.run(f"docker stack ps {core_stack_name}", shell=True, check=False)
-        # logs table like
-        #  ID                  NAME                  IMAGE                                      NODE                DESIRED STATE       CURRENT STATE                ERROR
-        # xbrhmaygtb76        simcore_sidecar.1     itisfoundation/sidecar:latest              crespo-wkstn        Running             Running 53 seconds ago
-        # zde7p8qdwk4j        simcore_rabbit.1      itisfoundation/rabbitmq:3.8.0-management   crespo-wkstn        Running             Running 59 seconds ago
-        # f2gxmhwq7hhk        simcore_postgres.1    postgres:10.10                             crespo-wkstn        Running             Running about a minute ago
-        # 1lh2hulxmc4q        simcore_director.1    itisfoundation/director:latest             crespo-wkstn        Running             Running 34 seconds ago
-        # ...
-
-    # TODO: find a more reliable way to list services in a stack
-    core_stack_services: List[Service] = [
-        service
-        for service in docker_client.services.list()
-        if service.name.startswith(f"{core_stack_name}_")
-    ]  # type: ignore
-
-    assert (
-        core_stack_services
-    ), f"Expected some services in core stack '{core_stack_name}'"
-
-    assert len(core_stack_compose["services"].keys()) == len(core_stack_services)
-
-    return core_stack_services
 
 
 SERVICES_AND_EXIT_CODES = [
@@ -90,6 +26,7 @@ SERVICES_AND_EXIT_CODES = [
     ("webserver", 0),
 ]
 
+MAX_TIME_TO_RESTART_SERVICE = 10
 
 # FIXME: https://github.com/ITISFoundation/osparc-simcore/issues/2407
 @pytest.mark.skip(
